refactor: 2 (manifest) — extract load_manifest + variant resolution to cli/manifest.py
Move load_manifest, resolve_options, resolve_variant_files and the variant
helpers (_select_variant, _apply_by_type, _dimension_entries, _collect_entries)
out of govkit.py into a dedicated cli/manifest.py that depends only on the path
kernel. govkit re-exports the 3 public functions, so callers and tests are
unchanged. govkit.py: 1198 -> 974 lines. No behavior change; 807 tests green.
